Refactored v8 value collection and reference

+++ cpp/jni/com_caoccao_javet_interop_V8Native.h
@@ -34,10 +34,10 @@ JNIEXPORT jobject JNICALL Java_com_caoccao_javet_interop_V8Native_execute
 /*
  * Class:     com_caoccao_javet_interop_V8Native
  * Method:    getLength
- * Signature: (JIJ)I
+ * Signature: (JJI)I
  */
 JNIEXPORT jint JNICALL Java_com_caoccao_javet_interop_V8Native_getLength
-  (JNIEnv *, jclass, jlong, jint, jlong);
+  (JNIEnv *, jclass, jlong, jlong, jint);
 
 /*
  * Class:     com_caoccao_javet_interop_V8Native
